+++ src/BaseOAuth.php
@@ -222,8 +222,10 @@ abstract class BaseOAuth extends BaseClient
     /**
      * Sends the given HTTP request, returning response data.
      * @param \yii\httpclient\Request $request HTTP request to be sent.
-     * @return array response data.
-     * @throws InvalidResponseException on invalid remote response.
+     * @return array|string|null response data.
+     * @throws ClientErrorResponseException on client error response codes.
+     * @throws InvalidResponseException on non-successful (other than client error) response codes.
+     * @throws \yii\httpclient\Exception
      * @since 2.1
      */
     protected function sendRequest($request)
@@ -231,11 +233,16 @@ abstract class BaseOAuth extends BaseClient
         $response = $request->send();
 
         if (!$response->getIsOk()) {
-            $statusCode = $response->getStatusCode();
-            throw new InvalidResponseException(
+            $statusCode = (int)$response->getStatusCode();
+            if ($statusCode >= 400 && $statusCode < 500) {
+                $exceptionClass = 'yii\\authclient\\ClientErrorResponseException';
+            } else {
+                $exceptionClass = 'yii\\authclient\\InvalidResponseException';
+            }
+            throw new $exceptionClass(
                 $response,
                 'Request failed with code: ' . $statusCode . ', message: ' . $response->getContent(),
-                (int) $statusCode
+                $statusCode
             );
         }