+++ src/ClientErrorResponseException.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * @link https://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license https://www.yiiframework.com/license/
+ */
+
+namespace yii\authclient;
+
+/**
+ * ClientErrorResponseException represents an exception caused by a "client error" server response status code (4xx).
+ *
+ * @since 2.2.15
+ */
+class ClientErrorResponseException extends InvalidResponseException
+{
+}